creates spqe files/ cleans up dead code.
