+++ src/qforte/ucc/__init__.py
@@ -1,5 +1,6 @@
 from .adaptvqe import *
 from .uccnvqe import *
-from .rsuccvp import *
+# from .rsuccvp import *
+from .spqe import *
 # from .uccvp import *
 from .uccpqe import *
